// src/auth/interface/AuthFormData.interface.ts
import { SubmitHandler } from "react-hook-form";


export interface ResponseUserAuthenticated{
    message:         string;
    user:            AuthFormData;
}


export interface AuthFormData {
    name:              string;
    email:             string;
    handle:            string;
    password:          string;
    confirmPassword?:  string;
    description?:      string;
    image?:            string;
    links:             string[];
}

export interface AuthFormProps{
    formType: "login"  | "register";
    handle?:  string;
    onSubmit: SubmitHandler<AuthFormData>;
}

// Generated by https://quicktype.io
// Generated by https://quicktype.io

export interface ResponseHTTPAxios {
    message: string;
    name:    string;
    stack:   string;
    config:  Config;
    code:    string;
    status:  number;
}

export interface Config {
    transitional:      Transitional;
    adapter:           string[];
    transformRequest:  null[];
    transformResponse: null[];
    timeout:           number;
    xsrfCookieName:    string;
    xsrfHeaderName:    string;
    maxContentLength:  number;
    maxBodyLength:     number;
    headers:           Headers;
    method:            string;
    url:               string;
    data:              string;
}


export interface Headers {
    Accept:         string;
    "Content-Type": string;
}

export interface Transitional {
    silentJSONParsing:   boolean;
    forcedJSONParsing:   boolean;
    clarifyTimeoutError: boolean;
}